requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
customtkinter==5.2.0
google-genai==0.3.0
python-dotenv==1.0.0
gtts==2.3.1
edge-tts==6.1.12
requests==2.31.0
beautifulsoup4==4.12.2
wikipedia==1.4.0
